refactor: memoized message rows
